feat: abstract utxo sweeper
