+++ src/components/Sweep.tsx
@@ -1,0 +1,186 @@
+import { Box, Button, Divider, Skeleton, Stack, Text as CText } from '@chakra-ui/react'
+import { type AccountId, type AssetId, fromAssetId } from '@shapeshiftoss/caip'
+import type { UtxoBaseAdapter, UtxoChainId } from '@shapeshiftoss/chain-adapters'
+import { bn, FeeDataKey } from '@shapeshiftoss/chain-adapters'
+import type { Utxo } from '@shapeshiftoss/unchained-client/src/generated/bitcoin'
+import { useCallback, useEffect, useState } from 'react'
+import { FaExchangeAlt } from 'react-icons/fa'
+import { useTranslate } from 'react-polyglot'
+import { Row } from 'components/Row/Row'
+import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
+import { usePoll } from 'hooks/usePoll/usePoll'
+import { useWallet } from 'hooks/useWallet/useWallet'
+import { bnOrZero } from 'lib/bignumber/bignumber'
+import { useGetEstimatedFeesQuery } from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
+import { selectAssetById } from 'state/slices/selectors'
+import { useAppSelector } from 'state/store'
+
+import { Amount } from './Amount/Amount'
+import { AssetIcon } from './AssetIcon'
+import { handleSend } from './Modals/Send/utils'
+
+const divider = <Divider />
+
+type SweepProps = {
+  assetId: AssetId
+  fromAddress: string | null
+  accountId: AccountId
+  onBack: () => void
+  onSweepSeen: () => void
+}
+
+const providerIcon = 'https://assets.coincap.io/assets/icons/rune@2x.png'
+
+export const Sweep = ({
+  assetId,
+  fromAddress,
+  accountId,
+  onBack: handleBack,
+  onSweepSeen: handleSwepSeen,
+}: SweepProps) => {
+  const [isSweepPending, setIsSweepPending] = useState(false)
+  const [txId, setTxId] = useState<string | null>(null)
+
+  const {
+    state: { wallet },
+  } = useWallet()
+  const translate = useTranslate()
+  const asset = useAppSelector(state => selectAssetById(state, assetId))
+
+  const { data: estimatedFeesData, isLoading: isEstimatedFeesDataLoading } =
+    useGetEstimatedFeesQuery({
+      cryptoAmount: '0',
+      assetId,
+      to: fromAddress ?? '',
+      sendMax: true,
+      accountId,
+      contractAddress: undefined,
+      enabled: true,
+    })
+
+  const handleSweep = useCallback(async () => {
+    if (!wallet) return
+
+    setIsSweepPending(true)
+
+    try {
+      if (!fromAddress)
+        throw new Error(`Cannot get from address for accountId: $accountIdcollateralAccountId}`)
+      if (!estimatedFeesData) throw new Error('Cannot get estimated fees')
+      const sendInput = {
+        accountId,
+        to: fromAddress,
+        input: fromAddress,
+        assetId,
+        from: '',
+        cryptoAmount: '0',
+        sendMax: true,
+        estimatedFees: estimatedFeesData.estimatedFees,
+        amountFieldError: '',
+        fiatAmount: '',
+        vanityAddress: '',
+        feeType: FeeDataKey.Fast,
+        fiatSymbol: '',
+      }
+
+      const txId = await handleSend({ wallet, sendInput })
+      setTxId(txId)
+    } catch (e) {
+      console.error(e)
+    }
+  }, [accountId, assetId, estimatedFeesData, fromAddress, wallet])
+
+  const adapter = getChainAdapterManager().get(fromAssetId(assetId).chainId)
+
+  const { poll } = usePoll()
+
+  useEffect(() => {
+    if (!adapter || !fromAddress) return
+    // Once we have a Txid, the Tx is in the mempool which is enough to broadcast the actual Tx
+    // but we still need to double check that the matching UTXO is seen to ensure coinselect gets fed the right UTXO data
+    if (!txId) return
+    ;(async () => {
+      await poll({
+        fn: () =>
+          (adapter as unknown as UtxoBaseAdapter<UtxoChainId>).getUtxos({ pubkey: fromAddress }),
+        validate: (utxos: Utxo[]) => utxos.some(utxo => utxo.txid === txId),
+        interval: 1000,
+        maxAttempts: 10,
+      })
+    })()
+  }, [adapter, fromAddress, poll, txId])
+
+  if (!asset) return null
+
+  return (
+    <Stack spacing={0} divider={divider}>
+      <Stack py={4} spacing={4} px={6} fontSize='sm' fontWeight='medium'>
+        <Stack flex={1} spacing={6} p={4} textAlign='center'>
+          <Stack
+            spacing={4}
+            direction='row'
+            alignItems='center'
+            justifyContent='center'
+            color='text.subtle'
+            pt={6}
+          >
+            {providerIcon && (
+              <>
+                <AssetIcon src={asset.icon} />
+                <FaExchangeAlt />
+                <AssetIcon src={providerIcon} size='md' />
+              </>
+            )}
+          </Stack>
+          <Stack>
+            <CText color='text.subtle'>
+              {translate('modals.send.consolidate.body', { asset: asset.name })}
+            </CText>
+          </Stack>
+          <Stack justifyContent='space-between'>
+            <Button
+              onClick={handleSweep}
+              disabled={isEstimatedFeesDataLoading || isSweepPending}
+              size='lg'
+              colorScheme={'blue'}
+              width='full'
+              data-test='utxo-sweep-button'
+              isLoading={isEstimatedFeesDataLoading || isSweepPending}
+              loadingText={translate('common.loadingText')}
+            >
+              {translate('modals.send.consolidate.consolidateFunds')}
+            </Button>
+            <Button
+              onClick={handleBack}
+              size='lg'
+              width='full'
+              colorScheme='gray'
+              isDisabled={false}
+            >
+              {translate('modals.approve.reject')}
+            </Button>
+          </Stack>
+        </Stack>
+        <Stack p={4}>
+          <Row>
+            <Row.Label>{translate('modals.approve.estimatedGas')}</Row.Label>
+            <Row.Value>
+              <Skeleton isLoaded={!isEstimatedFeesDataLoading}>
+                <Box textAlign='right'>
+                  <Amount.Fiat value={estimatedFeesData?.txFeeFiat ?? '0'} />
+                  <Amount.Crypto
+                    color='text.subtle'
+                    value={bnOrZero(estimatedFeesData?.txFeeCryptoBaseUnit)
+                      .div(bn(10).pow(asset?.precision ?? '0'))
+                      .toString()}
+                    symbol={asset.symbol}
+                  />
+                </Box>
+              </Skeleton>
+            </Row.Value>
+          </Row>
+        </Stack>
+      </Stack>
+    </Stack>
+  )
+}
